backend/pyproject.toml
[project]
name = "sgm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.5",
    "psycopg[binary,pool]>=3.2.3",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.0",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = []
lint = [
    "ruff>=0.8.0",
]

[tool.ruff]
src = ["app"]